making ssh app use any port
