+++ .gitignore
@@ -167,5 +167,6 @@ sqlite_storage/old_sqlite.db
 sqlite_storage/sqlite.db
 sqlite_storage/sqlite.db.bak
 
-ssh_hosts_app/old
+port_hosts_app/old
 .pre-commit-config.yaml
+src/sqlite_storage/sqlite.db
