+++ README.md
@@ -3,6 +3,6 @@
 
 ## project to store smaller coding projects
 To-Do:
-    - `ssh_hosts_app`: Get working in container and create k8s deployment
+    - `port_hosts_app`: Get working in container and create k8s deployment
     - `api_request`: Query SQLIte DB and render in browser
     - Use pre-commit for lint OR Use bazel to build lint and test
