+++ src/port_hosts_app/old/local_inventory.py
