+++ src/port_hosts_app/old/other_local_inventory.py
