+++ src/port_hosts_app/src/templates/index.html
@@ -1,26 +1,27 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <title>SSH Hosts</title>
+    <title>Open Port Hosts</title>
 </head>
 <body>
-    <h1>SSH Hosts</h1>
+    <h1>Open Port Hosts</h1>
+    <h1>Using Port: {{ port_servers[1] }}</h1>
     <table>
         <thead>
             <tr>
                 <th>Host IP</th>
-                <th>SSH State</th>
-                <th>SSH Server</th>
-                <th>Host Platform</th>
-                <th>Host OS</th>
+                <th>State</th>
+                <th>Server</th>
+                <th>Platform</th>
+                <th>OS</th>
             </tr>
         </thead>
         <tbody>
-            {% for ip_address, host_data in ssh_servers.items() %}
+            {% for ip_address, host_data in port_servers[0].items() %}
             <tr>
                 <td>{{ ip_address }}</td>
-                <td>{{ host_data.ssh_state }}</td>
-                <td>{{ host_data.ssh_server }}</td>
+                <td>{{ host_data.port_state }}</td>
+                <td>{{ host_data.port_server }}</td>
                 <td>{{ host_data.host_platform }}</td>
                 <td>{{ host_data.host_os }}</td>
             </tr>
