--- src/ssh_hosts_app/src/templates/index.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>SSH Hosts</title>
-</head>
-<body>
-    <h1>SSH Hosts</h1>
-    <table>
-        <thead>
-            <tr>
-                <th>Host IP</th>
-                <th>SSH State</th>
-                <th>SSH Server</th>
-                <th>Host Platform</th>
-                <th>Host OS</th>
-            </tr>
-        </thead>
-        <tbody>
-            {% for ip_address, host_data in ssh_servers.items() %}
-            <tr>
-                <td>{{ ip_address }}</td>
-                <td>{{ host_data.ssh_state }}</td>
-                <td>{{ host_data.ssh_server }}</td>
-                <td>{{ host_data.host_platform }}</td>
-                <td>{{ host_data.host_os }}</td>
-            </tr>
-            {% endfor %}
-        </tbody>
-    </table>
-</body>
-</html>
